refactor: move all filter related method to _FilterDataNode class
